+++ lizard_auth_client/models.py
@@ -1,10 +1,12 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
 
+from django.utils.encoding import python_2_unicode_compatible
 from django.conf import settings
 from django.db import models
 
 
+@python_2_unicode_compatible
 class Role(models.Model):
     unique_id = models.CharField(max_length=32, unique=True)
     code = models.CharField(max_length=255, null=False, blank=False)
@@ -14,7 +16,7 @@ class Role(models.Model):
 
     BILLING_ROLE_CODE = 'billing'
 
-    def __unicode__(self):
+    def __str__(self):
         return self.name
 
     @classmethod
@@ -52,6 +54,7 @@ class OrganisationManager(models.Manager):
         return self.get(unique_id=unique_id)
 
 
+@python_2_unicode_compatible
 class Organisation(models.Model):
     objects = OrganisationManager()
 
@@ -61,7 +64,7 @@ class Organisation(models.Model):
     name = models.CharField(max_length=255, null=False, blank=False)
     unique_id = models.CharField(max_length=32, unique=True)
 
-    def __unicode__(self):
+    def __str__(self):
         return self.name
 
     def natural_key(self):
@@ -85,6 +88,7 @@ class Organisation(models.Model):
         return org
 
 
+@python_2_unicode_compatible
 class UserOrganisationRole(models.Model):
     """Stores which roles in which organisations a user has."""
     user_model = getattr(settings, 'AUTH_USER_MODEL', 'auth.User')
@@ -92,7 +96,7 @@ class UserOrganisationRole(models.Model):
     organisation = models.ForeignKey(Organisation)
     role = models.ForeignKey(Role)
 
-    def __unicode__(self):
+    def __str__(self):
         return '%s %s %s' % (
             str(self.user), str(self.organisation), str(self.role))