+++ CHANGES.rst
@@ -2,10 +2,19 @@ Changelog of lizard-auth-client
 ===================================================
 
 
-1.10 (unreleased)
+1.11 (unreleased)
 -----------------
 
 - Added natural key to Organisation model.
+
+
+1.10 (2016-01-25)
+-----------------
+
+- Made sure the string representation on models also works on python 3 (it
+  also keeps working on python 2, of course). See
+  https://docs.djangoproject.com/en/1.8/ref/utils/#django.utils.encoding.python_2_unicode_compatible
+  [reinout]
 
 
 1.9 (2015-11-03)